Splitted EnsembleResult into Ensemble and EnsembleResult
