+++ maboss/ensemble/__init__.py
@@ -1,1 +1,2 @@
 from .result import EnsembleResult
+from .ensemble import Ensemble
